Update the ResponseWriter to implement various interfaces
- io.ReaderFrom - Allows io.Copy, etc. to use the
                  more optimized copying from
                  io.ReaderFrom that likely exists
                  on the wrapped ResponseWriter
- http.Pusher - Allows doing HTTP/2 server pushes
                if the wrapped ResponseWriter
                implements http.Pusher
- http.Flusher - Allows flushing buffered data in
                 the wrapped ResponseWriter if it
                 implements http.Flusher

+++ logger.go
@@ -42,7 +42,7 @@ func withLogger(serverLogger *slog.Logger) Middleware {
 				"response",
 				slog.String("duration", Duration(req).String()),
 				slog.Int("status", wrw.status),
-				slog.Int("length", wrw.length),
+				slog.Int64("length", wrw.length),
 				slog.String("type", wrw.Header().Get("Content-Type")),
 			)).Info("Request")
 		})
